Tornado

A dataset that contains the tornadoes and their characteristics in United States in 1950-2022.

Usage

tornados

Format

A data frame with 68,693 rows and 27 variables.

om

Tornado number. Effectively an ID for this tornado in this year.

yr

Year, 1950-2022.

mo

Month, 1-12.

dy

Day of the month, 1-31.

date

Date.

time

Time.

tz

Canonical tz database timezone.

datetime_utc

Date and time normalized to UTC.

st

Two-letter postal abbreviation for the state (DC = Washington, DC; PR = Puerto Rico; VI = Virgin Islands).

stf

State FIPS (Federal Information Processing Standards) number.

mag

Magnitude on the F scale (EF beginning in 2007). Some of these values are estimated (see fc).

inj

Number of injuries. When summing for state totals, use sn == 1

fat

Number of fatalities. When summing for state totals, use sn == 1

loss

Estimated property loss information in dollars. Prior to 1996, values were grouped into ranges. The reported number for such years is the maximum of its range.

slat

Starting latitude in decimal degrees.

slon

Starting longitude in decimal degrees.

elat

Ending latitude in decimal degrees.

elon

Ending longitude in decimal degrees.

len

Length in miles.

wid

Width in yards.

ns

Number of states affected by this tornado. 1, 2, or 3.

sn

State number for this row. 1 means the row contains the entire track information for this state, 0 means there is at least one more entry for this state for this tornado (om + yr).

f1

FIPS code for the 1st county.

f2

FIPS code for the 2nd county.

f3

FIPS code for the 3rd county.

f4

FIPS code for the 4th county.

fc

Was the mag column estimated?

Source

https://github.com/rfordatascience/tidytuesday/blob/master/data/2023/2023-05-16/tornados.csv